Add test to show incorrect handling of FlaggedApi on super method
Annotating a method with `@FlaggedApi` results in the overriding
methods behaving as if they were flagged with the same annotation. The
reason for that is because when reverting the addition of a new method
any of its overrides which are also added should be reverted too.

Unfortunately, that causes an issue when the overridden method is not
part of the same API as the overriding methods and no previously
released API is provided. In that case when processing the API
containing the overriding method Metalava will remove the method even
though the super class may not remove it (e.g. because it does provide
a previously released API which contains the method).

See b/379940628 for a concrete example.

This change adds a test to show the behavior with and without
previously released API, a follow-up change will fix it.

Bug: 379940628
Test: ./gradlew
Change-Id: I3bcc33f1a3858f1ee08c6761f5242ad4625185a0