Rename SdkVersion to ApiVersion
The AndroidX team requested this rename as it makes the code a little
confusing from their perspective as each AndroidX library has their own
version and does not really constitute an SDK. However, renaming it to
say `LibraryVersion` has similar issues with Android. This change uses
the generic `ApiVersion` instead.

`SdkVersion` was created from the AndroidX versioning scheme and was
designed to be a super set of the functionality required for both
Android (now and in future when it supports major.minor versions) and
AndroidX.

An alternative approach would be to have `ApiVersion` as a super type
and then have `SdkVersion` for Android and `LibraryVersion` for
AndroidX. That would allow `SdkVersion` to be simplified as Android
only uses the `ApiVersion.major` as a level (expanding to include
`minor` in future). However, that would lead to some duplication and/or
complication in the code and would not save any memory as the versions
are effectively memoized or used for a very short time.

Bug: 379646963
Test: ./gradlew
Change-Id: I742a572163259a348e1111b4311eedc4f2dcd4e9