Prevent creating SdkIdentifier with ANDROID_PLATFORM_SDK_ID
Previously, it was possible to create an `SdkIdentifier` with an `id`
of `ANDROID_PLATFORM_SDK_ID`, i.e. `0`. That was needed because it was
used in `DocAnalyzer.kt` but the preceding change removed that so this
change removes the possibility for doing that and removes unnecessary
checks elsewhere.

Rather than just check to make sure it is not `ANDROID_PLATFORM_SDK_ID`
this checks to make sure that the `id` is greater than or equal to `1`.

Bug: 380412656
Test: ./gradlew
Change-Id: I38527efae4069782903b213034ae640e99d4c2c7