Make Api implement ParentApiElement instead of ApiElement
Previously, `Api` needed to extend `ApiElement` so an `Api` instance
could be passed as the parent `ApiElement` in `ApiXmlPrinter`. However,
the preceding change added a `ParentApiElement` for use in
`ApiXmlPrinter` so `Api` no longer needs to extend `ApiElement`.

This change makes `Api` implement `ParentApiElement` directly. That
required that it added its own `update(SdkVersion)` method simply to
track `lastPresentIn`.

Bug: 379646963
Test: ./gradlew
Change-Id: I41dba95652bdda0e425f12547541ce0b37b66789