Update Api with Codebase version
Previously, `addApisFromCodebase(...)` would not call `Api.update(...)`
which meant it would not detect when classes were `@removed` from the
current source `Codebase` as was revealed by the test added in the
preceding change. This change fixes that and fixes the test.

This change also revealed a couple of unrealistic tests that would
request to add the current source but then provide sources which
did not include all the classes from the previous releases. That meant
those classes should have been treated as removed but were not. This
change then broke those tests. Those tests have been patched to make
them pass and TODOs added to ensure that they are fixed in future.

This does not affect AndroidX because they use `ApiJsonPrinter` and it
is only `ApiXmlPrinter` that cares about the versions set on `Api`. It
also does not break Android build as it provides sources that are at
least a superset of the previously released APIs and so will not treat
a class as removed unless it is tagged with `@removed`.

Bug: 379646963
Test: ./gradlew
Test: tools/metalava/scripts/gather-android-metalava-artifacts.py before
      # Apply this change
      tools/metalava/scripts/gather-android-metalava-artifacts.py after
      # Make sure that there are no differences.
Change-Id: I7b06a3e5a01ffcac2ad0070be4e5bb1be8bf616e