Maintain indentation in checkApiVersionsXmlContent(...)
Previously, it would discard all indentations in the expected and
actual content. That was because the actual content was indented using
`\t` characters which are not easy to use in a raw string as editors
can replace them with spaces. This change just replaces the `\t`
characters with 4 spaces before comparing which makes debugging test
failures and adding new tests easier.

Bug: 380412656
Test: ./gradlew
Change-Id: Ib567dc5b79a7276def1c8c3ccef7b5e021ba7341