Improve testing of historical backfill
Previously, the test would look for specific parts of the
`api-versions.xml` entry for `SdkExtensions`, directly by looking in
the XML output and indirectly through the `ApiLookup`. That meant that
some parts went untested. This change extracts the whole `<class>`
element for `SdkExtensions` and compares it against an expected
result.

That reveals a problem with the `getAllExtensionVersions()` method as
it should be reported as having been added as part of 31 but it is
reported as being added between 31 (S) and 32 (T). That will be fixed
in a follow-up change.

It tests in both system and public as the `SdkExtensions` appeared at
different versions in each API surface. There is currently one
discrepancy between system and public. System has:
```
<field name="AD_SERVICES" since="34" sdks="30:4"/>
```
Whereas public has:
```
<field name="AD_SERVICES" since="34" sdks="30:4,0:34"/>
```

It is not clear why they are different.

Bug: 381187230
Test: ./gradlew
Change-Id: If86cdf1736806702dfde73f8d7e4a4d80e9282db