Combine ApiElement.update(...) methods with default value
Previously, there were two separate methods `update(SdkVersion)` and
`update(SdkVersion, Boolean)`. The former called the latter passing in
a default value. That is a hold-over from when `ApiElement` was written
in Java. In Kotlin it is more idiomatic to have a single method with a
default parameter so this change makes that switch.

Bug: 379646963
Test: ./gradlew
Change-Id: Iedadfb26ed8f46e47e89739591202514e8a3cacb