Dedup creating Api from VersionedApis
Previously, `createApiFromAndroidJars(...)` and
`createApiFromVersionedJars(...)` would both create an `Api` object
and update it from a list of `VersionedApi`s. This change dedups that
functionality. It does that by pulling the call to `api.clean()` from
the latter into the calling method `generateFromVersionedApis(...)`
and then reusing `createApiFromVersionedJars(...)` in
`createApiFromAndroidJars(...)` which is then inlined as it is not
worth keeping as follow-up changes will eventually pull the creation of
the `VersionedApi`s list into `ApiLevelsGenerationOptions`.

Bug: 378479241
Test: ./gradlew
Change-Id: Iae06fae55861304ade8ca434ca652bf1a7983260