Generalize extracting information from matching file name
Replaces custom code for handling the `{version:level}` placeholder
with some general code for handling placeholders.

Adds `Property` to enumerate the properties that the placeholders
update. At the moment this just has `VERSION` but it will add more in
future, e.g. `SURFACE`, `MODULE`, etc.. These are referenced from
`Placeholder` and are responsible for taking a String extracted from
the file name, converting it into an appropriate type, checking to make
sure it is valid and if it is then storing it in the
`PatternFileState`.

Adds `Placeholder` to enumerate each of the different placeholders.
This is used in `parseParameterizedPattern(...)` to find the available
placeholders and create an appropriate regular expression that are
both then stored in a `PlaceholderPatternNode`.

Splits `ApiVersionPatternNode` into two parts. The `ApiVersion`
specific parts are moved into the `Property.VERSION` instance. The
general placeholder functionality goes in to a new
`PlaceholderPatternNode` class which takes a list  of `Placeholder`s
and a corresponding regular expression. The latter is used to determine
which files match and the former are used to extract the property
value.

Follow-up changes will add new `Property` and `Placeholder` instances.

Bug: 378479241
Test: ./gradlew
Change-Id: I53f42664e473b1eecee729e5df1fb904eda2141a