Generalize the handling of the VisibleForTesting annotation
The `VisibleForTesting` annotation is used to annotate an `Item` that
is publicly visible for testing purposes but should not be treated as
public for the purposes of generating an API surface. Previously, the
logic for handling that was specific to the Psi model. This change
generalizes that to use the `metalava-model` API and moves it to
`DefaultModifierList.createMutableModifiers(...)` so it is consistent
across all providers.

This does not fix Turbine because Turbine does not currently provide
access to constant symbols, e.g. `VisibleForTesting.PRIVATE`, but only
their constant value, e.g. `2`.

Bug: 383521289
Test: ./gradlew
Change-Id: I2fe7c0683c96a7a815776bf2b19a3a9bb4a7fef9