Allow any format version to be specified for --current-version
Previously, `--current-version` would only accept a single integer API
level. This change switches it (and `--first-version`) to support any
acceptable `ApiVersion`. It also moves the check that the version is
27 or more from the option to the `forAndroidConfig(...)` method as it
only applies when generating Android history.

A follow-up change will allow that to be used to specify the current
version when generating a JSON API history instead of tagging it on the
end of the `--api-version-names` option.

Bug: 388441272
Test: ./gradlew
Change-Id: I35b7f74212d7dc43320f81033a12fbd6a615d061