Add help information to FileFormat.Version
Previously, the help for `FileFormat.Version`s was hard coded in the
`HelpCommandTest` which meant that it was easy to get out of date. This
change moves the version specific help into the `FileFormat.Version`
instances ensuring that it is up to date. This revealed that there was
no documentation for `FileFormat.Version.V5`.

A side-effect of this change was some extra vertical white space being
added in the generated text. This will be fixed in a follow-up change.

Bug: 388809101
Test: ./gradlew
Change-Id: I865ff4721c4efc179f55661bc256e6c6c023ea55