Move help information to FileFormat.CustomizableProperty
Moves the existing help about the properties from `HelpCommand` to
the `FileFormat.CustomizableProperty` instances to which the help
applies. That reveals that most of the properties do not have help.
That will be rectified in future changes.

This slightly changed the help as the properties are in alphabetical
order but the previous help was not.

Bug: 388809101
Test: ./gradlew
Change-Id: Ic286375a09185e433ab747b5d436c12b27888c18