Prepare for improving the Config parsing
Previously, the `Config` and `ConfigParser` classes were added to allow
some configuration files to be passed to Metalava with the intent that
it was only temporary and would be replaced by something more useful.
Unfortunately, that work was delayed.

This change cleans up and prepares the existing code for follow-up
changes to improve the parsing. It improves documentation of
`config.xsd` and splits the tests into two parts, one
to test `--config-file` option and a lower level one to test the parser
and schema. It also adds new tests for multiple files. It was necessary
for the `ConfigFilesTest` to make `config` available from `Options`.

Follow-up changes will use `com.fasterxml.jackson.dataformat` to
perform the deserialization but as support for schema validation in the
underlying StAX parser may have problems it will continue to use the
`ConfigParser` to perform schema validation.

Bug: 391554590
Test: ./gradlew
Change-Id: Ice0b94bb9aa3c0f362fa9f18c22093a5fdd3379a