Parse config files using com.fasterxml.jackson.dataformat
Adds basic support for using `com.fasterxml.jackson.dataformat` to
parse configuration files into `Config` objects. It uses version
`2.15.0` of `jackson-dataformat-xml` and `jackson-module-kotlin` as
while they are not the latest they are available in
`prebuilts/androidx/external` and so should not require any additional
work by the AndroidX team. Some work will be needed in Android to add
it to the list of prebuilts used by Metalava.

`jackson-dataformat-xml` uses a StAX parser under the covers which has
limited support for validating using an XML Schema so this maintains
the existing `ConfigParser` validation to ensure that the XML being
parsed adheres to the schema.

As JacksonXML can serialize and deserialize this adds `roundTrip(...)`
to the `ConfigParserTest` to serialize and deserialize a `Config`
object. While Metalava does not use serialization itself it does help
during development when adding new classes to see how JacksonXML will
serialize them as that is also the input it will accept when
deserializing.

Bug: 391554590
Test: ./gradlew
Change-Id: I5007d0507419dcac716fa5c23c2d587ca8152725