Order ApiSurface by their order in ApiSurfaces
Previously, `ApiSurface` instances were unordered, this change orders
them based on their creation. order in `ApiSurfaces`. That is such that
an `ApiSurface` that is extended by another is always created before it
(otherwise the `extends` relationship could not be resolved).

Needed for the follow-up change to order signature files found while
scanning for historical signature files.

Bug: 378479241
Test: ./gradlew
Change-Id: Ifa774b676b4ab3d6aae8617fa3cb772803ddb884