Make VersionedSignatureApi take a list of files
This is needed for handling surfaces like the Android system API which
is split across multiple signature files, e.g. `public` and `system`.
To make the resulting object easier to debug it uses bash string
expansion syntax to share common prefixes and/or suffixes. That form
was chosen as it can be helpful to copy the value and use it in a shell
to easily find the files being loaded and view their contents.

Bug: 378479241
Test: ./gradlew
Change-Id: Icc474616c6de48b16b311b4ed4a5fab930a46cc3