Move creation of VersionedJarApi instance to a factory
Previously, creation of `VersionedJarApi` instances was hard-coded in
the `constructVersionedApisForHistoricalFiles(...)` and
`addVersionedExtensionApis(...)` methods. This change extracts that
into a `VersionedApiFactory` parameter passed to those two methods to
allow a follow-up change to create `VersionedSignatureApi` instances
instead.

The factory takes a list of `MatchedPatternFile`s as while
`VersionedJarApi` only takes a single `File`, `VersionedSignatureApi`
takes a list of them.

Bug: 378479241
Test: ./gradlew
Change-Id: I18f4bf6c1b54fd505fac5a8a2b44eb5fdaf57051