Add tests to verify that VersionedApis update Api consistently
An `Api` instance is updated with information about different versions
of the API in different forms, e.g. source code, signature files and
jar files. It relies on the same version of each class to update `Api`
in the same way irrespective of which form it is in.

Unfortunately, that is not true, at least for interface classes so this
change adds tests to show the behavior for normal classes, interfaces,
enums and annotation classes. They were chosen because they have
special handling in `addApisFromCodebase(...)`. It shows that the
inconsistent handling of interfaces leads to differences in behavior
depending on the ordering.

Bug: 394259916
Test: ./gradlew
Change-Id: I6e7284da8dc20962a1004fb41b242e03ab996f3a