Remove dead code left over from b/312558482
Historically, a `ClassItem` for an interface would store the first
entry in the `implements` list into the `superClassType` property. That
was an implementation detail of the underlying `PsiClass` that leaked
out. As a result there was code in `addApisFromCodebase(...)` which
would treat the `superClassType` for an interface class as being part
of the `implements` list.

The `ClassItem` behavior was fixed in b/312558482 so that the
`superClassType` of an interface class was always `null` resuling in
the code in `addApisFromCodebase(...)` becoming dead code which this
change removes.

Bug: 394259916
Test: ./gradlew
Change-Id: If21dc2667ad365ce824ede5d89d791b131c3171b