Fix unqualified type issues
Unqualified types are assumed to be part of the `java.lang` package,
however that is not always true. Follow-up changes will improve the
handling of unqualified types to check whether the class exists in the
`java.lang` package. They reveal a number of issues with existing
tests using unqualified types incorrectly. This change fixes them.

Errors include:
* Failing to qualify a type, e.g. `Other` instead of `test.pkg.Other`.
* Using `Int` instead of `int`.
* Incorrect ordering of parameters, e.g. `str: String` instead of
  `String str` which treats `str` as a type.
* Additional `implements` keyword which ends up being treated as a
  type.
* Missing type parameter.

Bug: 394789173
Test: ./gradlew
Change-Id: Ib2040cb51cfbe8ab9ea2c5cff0971adc73cec47a