Require android.txt files in android-jars-to-signature
Originally, `android-jars-to-signature` was created to generate
`android.txt` files from `android.jar` files for versions that did not
have an `android.txt` file and also update existing `android.txt` files
to add missing information that was not present before. However, now
every version has both `android.jar` and corresponding `android.txt`
files there is little point in having it work without `android.txt`
files so this change makes that a requirement.

If there is ever a need to use it to create a completely new
`android.txt` file then that can easily be dealt with by creating an
empty file.

Bug: 396431439
Test: ./gradlew
Change-Id: I4c9da8a8bcc2b023babf521c5bf33d892dfc906f