Add test to generate system API from jar file
When the Android build generates the main `api-versions.xml` files for
the `public`, `system`, etc. APIs it always passed in jars containing
stubs, i.e. a pre-filtered API, and so it does not need to pass in the
normal `--show*` options for the surface.

As a result the consistency check between the command line options and
the API surface configuration fails when generating the `system` (or
similar non-`public` surface) as that expects a `--show*-annotation`
option to be specified but they are not.

This change adds a test to show that failing behavior. A follow-up
change will fix that.

Bug: 378479241
Test: ./gradlew
Change-Id: Ibc9bffb5e20774d725fe4a34e2f5da73ac8a64a8