Make it easier to add Files and other objects to args list
Previously, it was necessary for a `File` to be converted to a `String`
before being added to the `args` list. That was whether it was added
directly or was part of another list that was added. This change adds
some `plusAssign` operators which will allow any object, collection or
array of objects to be added to the `args` list by converting them to
`String`s first.

Bug: 396431439
Test: ./gradlew
Change-Id: I62cdd37ee912552532f322bd069fda5c5b2d2176