Add test for resolving a delta with annotation values
Resolving a delta between two identical `Codebase`s should result in
an empty file but that is not the case when one `Codebase` is from the
Psi model, the other is from a signature file and there are annotations
with attributes used. That is because comparing `AnnotationItem`s
across `Codebase`s from different models does not work because the
attribute value do not compare equal, e.g. if the attribute value is an
enum constant then the text model will represent it as a `String` but
if loaded from a jar file the Psi model would represent it as a
`com.intellij.psi.impl.compiled.ClsEnumConstantImpl`.

This change adds a test that shows this behavior. Note, it is not
possible to test this in `SnapshotDeltaMakerTest` as it requires Psi
and text `Codebase`s and the former is not accessible from that test.

A follow-up change will fix the test.

Bug: 394597358
Test: ./gradlew
Change-Id: I9c0016cdc3f8ab8ee91c817a5a87290510353691