Support supplying a class path when loading a jar
Loading a `Codebase` from a jar that is not self-contained, e.g. a test
jar, can lead to differences in behavior than when loading one from a
self-contained jar, e.g. `android.jar`. That is because the former can
have unresolvable classes. This allows a classpath to be provided
similar to when creating a `Codebase` from sources.

Bug: 397882602
Test: ./gradlew
Change-Id: Ieaf8584069f02ad1774bc026f6f08e5c7eb917e3