Fix signature tests that depend on SUPPORT_TYPE_USE_ANNOTATIONS
Previously, some tests depended on `SUPPORT_TYPE_USE_ANNOTATIONS` to
select different output based on the setting of that property.
Unfortunately, the expected output when it is `true` is generally wrong
as it was written on the understanding that the type annotations would
be added to format version 2.0 but that is not accurate. Type use
annotations (including nullability annotations) are only written out to
signature files when using a `FileFormat` that has:
```
    kotlinNameTypeOrder = true,
    includeTypeUseAnnotations = true,
    kotlinStyleNulls = false,
```

This change adds a `TYPE_USE_FORMAT` and uses it for those tests,
replacing the `SUPPORT_TYPE_USE_ANNOTATIONS` conditional that selected
the expected signature output with fixed content.

Bug: 318411790
Test: ./gradlew
Change-Id: I619b2dc07a7f4fe5d4d0e9cd3ea6eb5fabf2dd41