Make BaseConfigParserTest.roundTrip(...) more realistic
Previously, it would call `XmlMapper` directly to read and write
strings and did not actually use `ConfigParser` to read the contents as
that can only read `File`s. This makes the method more realistic by
reading and writing to a `File`. It also adds an extension function to
`Config` to write to a `File` which will be used in follow-up changes.

Bug: 399846595
Test: ./gradlew
Change-Id: Ia2d7f5a9cea4581b9c1a844c9c818c773592ec01