Add support for caching TestFiles
Some `TestFile`s, e.g. `JarFromSourcesFile` are expensive to create. In
order to amortize the cost it is helpful to reuse them across multiple
tests. This change adds `TestFileCache` and `TestFileCacheRule` to
support that.

It is used in a follow-up change to test the behavior of annotation
attributes on annotations loaded from jar files. Using the cache
gives about a 5x speed up on the test class run time.

Bug: 400664604
Test: ./gradlew
Change-Id: I12fb4e68d570bef893c8bea629a01c1530c45ed4