Ensure output writer is flushed and closed
Change https://r.android.com/3494670 added support to `signature-cat`
to write to a file. As part of that it flushed after writing out a
package to try and ensure that the file was written out correctly.
However, it did not work when the `Codebase` was empty, in that
case the header would be written to the buffer but not flushed so
the resulting file would be empty.

This change removes the call to `flush()` as that just obscures the
real issue which is that the file is not closed (which also flushes).
Instead it makes sure that the file is closed. It will also close
`stdout` but that does not cause any harm.

Bug: 396206005
Test: ./gradlew
Change-Id: I469716c0ba33463c02f30015a6799ebb6bba3786