Align how ensureCompanionJvmStatic and ensureCompanionJvmField checks if a class is a companion
One used to check if a class had the companion modifier, and the other checked if the class name was "Companion". Switch to checking both, since the goal of the check is to provide a better interop experience for Java developers by not requiring they use the Companion class, but named companions are often used for builders where it makes sense for the Java client to use the inner class name.

Bug: 401569415
Test: KotlinInteropChecksTest
Change-Id: I1426e01afa46c346426cb5646d1f49e2212ef60f