Compose API Guidelines
This is a first pass at the Compose API guidelines. It is based on the currently repo published guidelines, with a big focus on shortening as I copied.

I'm still adding stuff - but I wanted to get them out there so that people can start reviewing the first few sections and add more as further CLs. That way it doesn't become a MEGA CL.

Adding alanv for general review of file structure, setup, etc.

Preview here: go/androidx-compose-api-guidelines ?  cl=683274432

PiperOrigin-RevId: 735566646
Change-Id: If373fd054adb43f0d4fbfb8166033460ade6f548