Add value examples for all primitive types
Maps the inconsistencies in the handling of different primitive values.
Most of them have some inconsistencies. The worst are special values
for float and double.

Working on `ValueExample`s was a little tricky as it was not possible
to just run one, so this added a mechanism to temporarily limit any
tests that depend on it to a select few, usually one.

This also highlights the differences between Turbine and Psi.

Bug: 354633349
Test: ./gradlew
Change-Id: Iaacc805398434f7da297e0e7a0807f9a48e4e8f6