Verify that non constant values cannot be used in fields
Previously, a non-constant value, e.g. a string array, could have an
expectation other than `NO_INITIAL_FIELD_VALUE` but that would be wrong
as a field cannot have an initial value that is not a constant.

This adds a check based on the `ValueExample.javaType` to detect fields
that are not constants and verify that have `NO_INITIAL_FIELD_VALUE`.

Bug: 354633349
Test: ./gradlew
Change-Id: Ifb8ff9d6d76b21ad4f8be2b772d1ed28405d896a