Use generics in ParameterizedRunner
Previously, it was difficult to determine what the types of the
arguments were in `ParameterizedRunner` because it used `Any` as the
type. That was a holdover from when `CustomizableParameterizedRunner`
used reflection on top of `org.junit.runners.Parameterized` but that
is no longer necessary for our custom implementation.

This change adds a type parameter for the argument. In the process
the unnecessary wrapping of the argument in an array only to then
explode it immediately when formatting was removed.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia0ec9341effeeab1ca0ccbec96add23f0f0e946c