Add testing of values read from signature files
There are lots of failures which have been baselined. The failures fall
into two basic categories:
1. Passing the input through as a `String` instead of parsing it into
   some more appropriate representation.
2. Simple formatting inconsistencies due to just using the input string
   representation.

They will eventually be fixed as part of this work.

Note: This does not have different expectations for signature files as
it does for Kotlin because signature files are not written by hand,
except in tests. So, the values in the signature files are the values
written from either Kotlin or Java and so covered by the existing
tests.

Bug: 354633349
Test: ./gradlew
Change-Id: I854dc71a63ef573784c24e133539fc8649f5c2d6