Support different expectations for Kotlin
The main purpose of the test is to map the existing Psi behavior for
both Java and Kotlin to ensure that switching to a new `Value` model
does not break existing code. In that respect it makes sense to make
sure that the Kotlin tests pass for Psi just like the Java tests.

This change adds a set of Kotlin expectations that default to the same
as the Java ones but can be tweaked for specific `ValueExample`s. They
will be used in the follow-up change that adds tests for Kotlin.

Bug: 354633349
Test: ./gradlew
Change-Id: I486cff78bfb3d4e18f4fa73826d7994c24645075