Add PsiValueFactory and use for all value use sites
Previously, the Psi model implementation just used the
`ValueProvider.UNSUPPORTED` and `OptionalValueProvider.NO_VALUE`
placeholders. This change adds `PsiValueFactory` to create
`ValueProvider`s which delegate to itself to create `Value`s from the
`UExpression` and `PsiAnnotationMemberValue` hierarchies and uses
that for field, annotations and default attribute values.

Currently, it just throws a `ValueProviderException` but follow-up
changes will add support for the different types of `Value`s.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia97e046d0b6738654f70195cafaee29cb698e7b6