Add LiteralValue hierarchy
Expands the `Value` model by adding support for literal values, i.e.
those values which have a simple literal representation in the sources,
e.g. ints, longs, chars, booleans, Strings, etc. That does not include
more complex values like class literals and array literals.

As part of that it adds the `ValueFactory` which will be used for
creating instances of `Value`s and ensuring that they are consistently
created, e.g. `createLiteralValue(TypeItem?, Any)` will perform type
conversions on numbers if necessary to ensure they conform to the
`TypeItem`. If no `TypeItem` is provided then it will just select the
appropriate `LiteralValue` subclass to use based on the type of the
`underlyingValue`.

It also adds `ValueKind` as that is useful for parameterization of
various bits of code, e.g. tests. `Value.kind` implementations add no
overhead to each instance.

Bug: 354633349
Test: ./gradlew
Change-Id: I7611d5bbccbec50b87500067006390104a60bd51