Add ValueStringConfiguration.unwrapSingleArrayElement
When writing an annotation attribute value that is an array containing
a single element it can be written without the enclosing braces, e.g.
instead of `{"single"}` it can be `"single"`. That form is often used
when writing out annotation attributes. That change adds the
`unwrapSingleArrayElement` option to `ValueStringConfiguration` to
allow the caller to select that behavior.

Bug: 354633349
Test: ./gradlew
Change-Id: I4449d89386c6b041bed33cfeacadd382f3327383