Support conversion of integer types to char
Previously, integer types (i.e. `byte`, `int`, `long`, `short`) could
not be converted to a `char`. However, that is needed when parsing
signature files as sometimes a char is written out as its numeric code.
This change supports converting from those integer types to `char` and
updates the lossy conversion checking to support chars as well as
`Number`'s.

Bug: 354633349
Test: ./gradlew
Change-Id: If27c7a824f357ca70c5e5ec6a0f3670483963ba9