Support parsing LiteralValues in ApiFile
Makes use of `ValueParser` for `FieldItem.initialValue` and
`MethodItem.defaultValue`. The preceding change does the same for
`AnnotationItem`s created by `ApiFile`.

The `Value` tests are quite lax and set up to fail only when a non-null
`Value` is provided and does not match the expected `Value` (or a
non-`ValueProviderException` exception is thrown). Otherwise, they
throw an `AssumptionViolatedException` which causes the tests to be
recorded as ignored. That meant it was necessary to manually verify
that the literal tests are not ignored on Turbine. They are not and
they pass.

Removes `OptionalValueProvider.NO_VALUE` as it is no longer used.

Bug: 354633349
Test: ./gradlew
Change-Id: I56ff5ce4f3247330e7a32f166e7c1777f73bf399