Qualify annotation in generateAnnotatedTestClass(...) for signatures
Signature files require annotations that are not known to Metalava to
be fully qualified, otherwise they are assumed to be in the
`androidx.annotation` package. This change adds the package to the
annotation used in `generateAnnotatedTestClass(...)` as otherwise it
will fail to resolve properly which will cause problems when trying to
handle some `ValueExample`s. At the moment no tests rely on that but
follow-up changes will.

Bug: 354633349
Test: ./gradlew
Change-Id: I3b3ae35ce8546c1ee03638c4d255c10c7d23a16a