Add testLegacyValue() to CommonParameterizedFieldValueTest
Previously, the value returned by `FieldItem.initialValue(Boolean)` was
tested by comparing its `toString()` value but that was lossy and not
representative of how it is used. This change tests the value
explicitly, adding `expectedLegacyValue` / `expectedKotlinLegacyValue`
to `ValueExample` to set the expecttations. They mirror the existing
properties `expectedLegacySource` / `expectedKotlinLegacySource`
respectively.

This only adds expectations for those `ValueExample`s for which
`FieldItem.initialValue(true)` will return a non-null value. Follow-up
changes will add more as needed.

As is done for the other expectations the behavior of the Psi model is
taken as the standard and any deviations from that by other models are
baselined, even when a model's behavior makes more sense.

Bug: 354633349
Test: ./gradlew
Change-Id: Ib64b58bfd72d68f3ac96b8426a86275ef38d3dad