Separate handling of unqualified classes from TypeItemParser
Previously, `TypeItemParser` hard-coded the handling of unqualified
classes by treating it as being from `java.lang` package if it
contained a class matching the unqualified name, or leaving it as
unqualified and reporting an error. This separates that behavior out in
to a new `UnqualifiedClassHandler` type.

This is in preparation for using the `TypeItemParser` to parse class
literal references in signature files. It is needed for that as there
are lots of unqualified class references which need to be handled in a
different way, e.g. resolving against the containing package.

Bug: 354633349
Test: ./gradlew
Change-Id: I61271cdf846df90b153004a69916c9150b9b2130