Clean up errorReporter in TypeItemParser
Previously, the non-caching parts of `TextTypeItemParser` were
separated out into `TypeItemParser` but there was a little nastiness in
the handling of the `errorReporter` as `TypeItemParser` had to make it
`open` so that `TextTypeItemParser` could override it in order to
intercept the reported errors so it could make sure not to cache types
with errors as that would prevent other errors for the same type to be
reported elsewhere.

This change cleans that up by adding a secondary constructor which
wraps the `errorReporter` supplied in a `CountingErrorReporter` which
maintains the `errorCount`.

Bug: 354633349
Test: ./gradlew
Change-Id: Icbfe6d21d014c03e027d0a27e341e6e157c1588a