Add stubs tests for field values
There is a flaw in how Metalava handles `static final` fields with
unknown values and with `Class` values. This change adds some tests to
show that. It provides examples for all primitive types, `String`s and
a general class type (in this case `Runnable`) as each type has its
own default value.

Bug: 407951018
Test: ./gradlew
Change-Id: I5dae33d49f2d99db861ff95e8076c7152f5a9557