Use Value in SdkFileWriter
Switches `SdkFileWriter` from using the legacy values to `Value`. This
has the side effect of making it work on Turbine too.

This uses `Value.asString()` as the `SdkFileWriter` needs the string
value of a field, even when it is defined in terms of another field.
That is because in Android build some `@SdkConstant` annotated fields
reference other fields, e.g.
https://source.corp.google.com/h/googleplex-android/platform/superproject/main/+/main:frameworks/base/core/java/android/os/PowerManager.java;l=3347;drc=f50d2a29ac8af8f2ac55d974e324911e30a3b06b

Bug: 354633349
Test: ./gradlew
      # Before applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py before
      # After applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py after
      # Make sure that there are no differences.
      meld before after
Change-Id: Ia9fc113358ecc3da689ae9b25d43aae70a722a48