Refactor handling of field expectations
Previously, the expectations for fields of non-constant types were set
to `null` by having `constantFieldLegacySourceExpectation` (which sets
expectations for fields to `null`) fallback to the `ValueExample`'s
expectations. However, that would not work for `Value`s as that sets
the expectations of some constant types to be `ConstantFieldValue`s.

This change replaces that mechanism with a new
`TransformFieldExpectation` that applies a transform to the expected
values for fields. This change uses it for handling
`expectedLegacySource` and `expectedLegacyValue` and a follow-up
change will use it for `expectedValue`.

Bug: 354633349
Test: ./gradlew
Change-Id: I9b296183beceab615b0cf26f9e452f7160e396ec