Change default toValueString(...) for special floating values
Previously, it would just use `NaN`, `-Infinity` and `Infinity` for
the corresponding special floating point values. The problems with that
are:
* The same representation is used for floats and doubles.
* It differs from the rest of the `toValueString(...)` implementations
  which use a source representation of the value.

This changes it to use a source representation for the value. It chose
the values currently used by `CodePrinter` as they are compact and
clearly differentiate between floats and doubles.

Bug: 354633349
Test: ./gradlew
Change-Id: I9880ec6b7b1a5e5c915f111fd47ca43f91e79f67