UAnnotation is not a UExpression
Previously, a few places checked to see if a `UExpression` was a
`UAnnotation`. However, the latter is a `UElement`, which is a super
type of `UExpression`. While it is possible for a class to implement
both `UAnnotation` and `UExpression` there do not seem to be any such
classes in the current Psi libraries. There may have been some in the
past though.

This change cleans up that code as it is confusing to have code that is
never run and makes it appear that it supports annotations when it does
not. If the analysis is wrong and the branch could be taken then just
removing it entirely could lead to silent problems that will only be
found later. To avoid that this change keeps the branches but makes
them throw an exception making 100% sure that they are never taken.

Once it has been verified that the branches are never taken in actual
use (not just in Metalava tests) then they will be removed.

The `ExtractAnnotations` had another branch (immediately after the
`is UAnnotation` branch removed by this change) which handled
`UCallExpression`. That appears to be the branch which actually handled
nested annotations as in Kotlin nested annotations looks like normal
class construction and are represented by a call to a constructor. So,
this change moved the comment from the `is UAnnotation` branch into the
`is UCallExpression` branch so it is not lost when the first branch is
removed in future.

Bug: 354633349
Test: ./gradlew
Change-Id: Ifd47a4e610d754e8bb6fc516a689d19211ea656c