Add ValueStringConfiguration.nestedValueAppender
Previously, `Value.toValueString(...)` would always call itself on
nested `Value`s which meant that the caller had no way to affect the
formatting of those nested `Value`s, other than whatever control was
provided through the `ValueStringConfiguration`. This change adds
`nestedValueAppender` to the `ValueStringConfiguration` to fix that.
It provides a function that is used to format the nested `Value`.

This change uses that to ensure that `ArrayValue.debugStringForValue()`
uses the debug string representation of its nested `Value`s. Follow-up
changes will use that for annotations too.

Bug: 354633349
Test: ./gradlew
Change-Id: I8a0408edc76934074fb046af970fbb95e3bdc97e