Include attributes in PsiMethodItem.legacyDefaultValue
Previously, `CodePrinter.toSourceExpression(...)` (which is only used
for `PsiMethodItem.legacyDefaultValue`) would just include
`@<qualified-name>` for a `PsiAnnotation`. While it does not seem to
affect any signature file in Android or AndroidX this change fixes it
to improve consistency with other legacy value uses. That makes it
easier to add new tests and will make it simpler to migrate to `Value`
in future.

Bug: 354633349
Test: ./gradlew
      # Before applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py before
      # After applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py after
      # Make sure that there are no differences.
      meld before after
Change-Id: I53ec9e832c6c44d3c4d9d68188102b4b243e7ec0