Dedup scanning for closing quotes of a string literal
Previously, `getToken(...)` had two blocks of code for scanning over
string literals, one of which was a simple state machine that checked
for newlines in the string and the other was simpler but did not check
for newlines.

This change combines the two resulting in `scanForClosingQuotes()`
which is a simple scanner which checks for newline characters.

Bug: 354633349
Test: ./gradlew
Change-Id: I80e999fa3d76094e0e58d3c9bf4bf58ea429ae87