Prepare scanForEndOfToken(...) for balance parentheses
Previously, `scanForEndOfToken(...)` would only balance opening and
closing of type argument/parameter lists within `<...>`. It used an
`incompleteDepth` count to track the number of `<` characters that
had been seen without a matching `>` character. That would not work
for tracking a mixture of `<...>` and `(...)` blocks as it could not
differentiate between `<...>` and `<...)`. This change refactors it
to allow support for `(...)` blocks.

Instead of using a single count within a method it uses recursion, to
track the unbalanced `<` characters on the stack. It does that by
splitting the method into two, one for scanning the whole token and
one for scanning a token fragment, i.e. something between `<...>`.
The code is almost identical (only differs in its handling of the end
of the token) so to avoid duplication and performance hit it adds a
single common inlined method that is used in both.

Bug: 354633349
Test: ./gradlew
Change-Id: I615e91bcb745fa00071450b408e4cb2f4f739307