Add test for using int field constant as long
Shows the behavior of `long value = INT_CONSTANT`. The expected value
is wrong. The field reference's value should be a long which was
originally specified as an int as that is the value it would be if the
field reference was inlined.

That will be fixed in a follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: Ic0e59c1cc7dbc7617c5825dfd10f43751e447fb6