Only write field values for static final fields
Previously, `FieldItem.writeValueWithSemicolon(...)` would write out a
field value for non-`static final` fields. That was unnecessary as
signature files do not contain them and stubs do not need them. In
stubs, non-final fields do not need an initializer. Final but
non-static fields cannot exist in interfaces (or annotations), they can
only exist in a class (or enum) and they can just use an initializer
block.

This change makes that explicit. It did break some tests in
`ParameterizedFlaggedApiTest` as it switched from using a field value
to using a separate initializer block. It also broke a test in
`ApiFileTest` where it removed a value for a field from a signature
file.

Bug: 354633349
Test: ./gradlew
Change-Id: Ibb7668d4f820fdb834074e55318adada9c10953c