Remove FieldItem.legacyInitialValue()
Previously, `legacyInitialValue()` was only used in tests. Removing it
and all those uses does not reduce code coverage as the tests all have
parallel tests for `constantValue`.

One of those tests was broken on Turbine due to differences in the
behavior of `legacyInitalValue()` so removing that from the test fixed
it.

Bug: 354633349
Test: ./gradlew
Change-Id: Icfe6bbe7bc38e657adf448c544f054c5588a8d7e