Stop including default values in ANNOTATION_TO_SOURCE
Previously, `LegacyValueUseSite.ANNOTATION_To_SOURCE` included default
values in its legacy source output as it was mistakenly believed that
was needed for some output files, e.g. signatures or stubs. However,
the previous change revealed that it is only needed in
`AnnotationFilterEntry.fromAnnotationItem(...)`.

This change removes it and updates the baseline files for Turbine and
Text which both rely on the `DefaultAnnotationItem.toSource(...)` which
does not yet support `showDefaultAttrs = true`. As a result the tests
failed previously as they used `showDefaultAttrs = true` but now the
tests do not use that they pass on Turbine and Text.

Bug: 354633349
Test: ./gradlew
Change-Id: I5367921c349f48ffefc1e5514dbb37279750376e