Use interface class name in DefaultValue.appendToStringTo(...)
Previously, `DefaultValue.appendToStringTo(...)` used the simple name
of the implementation class to indicate the type of the value. That
meant that different implementations would have different string
representations and `assertValuesAreStrictlyEqual(...)` would fail.
That makes it difficult to provide alternate but equivalent
implementations.

This change adds the `valueKClass` to `ValueKind` and uses its
simple name. That ensures that two implementations that return the
same `ValueKind` can compare equal.

Bug: 354633349
Test: ./gradlew
Change-Id: Ide5850f6b0eb00901d14cceddd214e092567e095