Allow ArrayValue to hold one kind plus fields
Previously, `createArrayValue(...)` would fail if it was supplied with
two kinds of values, even if one of those was a field that could be of
the same kind as the non-field. This change allows for fields.

The check was added when it was thought that the kind of a field was
known at `ArrayValue` creation time but that is not true. Longer term
it would make sense to add a check when consuming the values, e.g. in
`asFlatList()` or maybe a completely new method.

Bug: 354633349
Test: ./gradlew
Change-Id: I6364f7ba3e60cf7f468c71be7c056fe41de2c5e0