Add LazyFieldReferenceValue which lazily gets its constant value
Previously, Psi and Text would resolve a `FieldItem` before creating a
`FieldReferenceValue`. That caused problems if the value was used
during `Codebase` construction as it could resolve a class before it
has been created. The reason for resolving the `FieldItem` was to
retrieve its `FieldItem.constantValue`.

This change adds a new `LazyFieldReferenceValue` which defers the
resolving of the `FieldItem` until `asLiteralValue()` is called.

If it turns out `asLiteralValue()` needs to be called in Psi during
`Codebase` construction than the constant value will need to be
computed from Psi specific objects. However, even if that is needed
the Text model will still need to use this so this is not wasted.

Bug: 354633349
Test: ./gradlew
Change-Id: I17742f4416835e5f7827a41a8488a71daa04bc37