Move getRetention() to AnnotationClass.retention
Annotation classes are special in that they have class specific data
that needs to be cached, e.g. retention policy. Previously, that was
cached within the `ClassItem`. This change adds a separate
`AnnotationClass` interface which is cached in `ClassItem` for
annotation classes and encapsulates the retention.

Follow-up changes will add more information to the class, e.g.
annotation attribute defaults.

Bug: 354633349
Test: ./gradlew
Change-Id: I1a603d75205e1f475daca59bf2f15682c0ea1d88