Separate filtering/printing in writeAnnotation(...)
Previously, `ExtractAnnotations.writeAnnotation(...)` mixed filtering
and printing of attributes and values. It used `CodePrinter` to print
values and it would sometimes filter the value out. That complicated
the logic and required special support from the `StringPrintWriter`
which could drop content that had already been written to it.

This change separates the filtering and printing to simplify the logic
and future changes.

Bug: 331752084
Test: ./gradlew
Change-Id: I0150147636cf05a582ad6520187afa56ecc6c16a