Replace AnnotationAttribute.legacyValue with value
Previously, `AnnotationAttribute.legacyValue` was only used in tests.
This change replaces them with uses of `AnnotationAttribute.value` to
ensure continued test coverage.

It fixes a couple of `CommonAnnotationItemTest`s for the text model. It
also fixes `annotation with enum values` tests for Psi, previously they
returned an object of a Psi specific class. It does not fix those tests
for the snapshot as `Psi/UAnnotationItem` do not implement
`AnnotationItem.snapshot(...)` correctly. That is a known issue and is
one of the things that this work is intended to fix.

Bug: 354633349
Test: ./gradlew
Change-Id: I4e0c28fd4a66a436ebbb7fdd72f5a0985f7d5346