Simplify mixing of inline tags and text
Previously, the grammar was a little restrictive on how inline tags
and text could be mixed, e.g. it would not allow an inline tag to be
separated from the start of the line by a space, i.e. it would allow:
```
 *{@code ...}
```

But not:
```
 * {@code ...}
```

This change simplifies the mixing of inline tags and text to address
that.

Bug: 429965593
Test: ./gradlew
Change-Id: I55bb5ce80f4da624fad146d884da64efe157e092