Remove unsupported cast from test signature file
Previously, Metalava could generate casts in a signature file and so it
was important that a test verify that they can be read from signature
files. However, Metalava can no longer generate casts into signature
files, and no existing signature files include casts so this change
fixes the test in preparation for future changes which will fail if
they encounter a cast expression.

Bug: 354633349
Test: ./gradlew
Change-Id: I2fadd073b190348e5e08dfff0eab8ceac6693478