Use default AnnotationItem.toSource() for UAnnotationItem
Previously, `UAnnotationItem` had its own implementation of
`AnnotationItem.toSource()`. This change removes that so it will use
the default implementation.

As `UAnnotationItem` formatted values slightly differently to
`PsiAnnotationItem` it was necessary to add a Kotlin specific
`Settings` for `LegacyValueFormatter.ANNOTATION_SOURCE_FORMATTER`
(which is used for the default `toSource()` implementation).

Even with those `Settings` the behavior of the default `toSource()` is
not identical to the `UAnnotationItem.toSource()` for all possible
values. However, it is the same for all values used in AndroidX
signature files which are the most important ones.

Where there were differences that broke a test the test was changed.

Bug: 354633349
Test: ./gradlew
      # Tested on Android and AndroidX
Change-Id: I4354ccbf6354859ca28c848d3366c1a9c226e122