Add test to show behavior of unresolvable flag field
Previously, change https://r.android.com/3605699 changed the behavior
of an unresolvable flag field and it was not caught by the tests so
this change adds a test to cover that. The current behavior is
probably more correct than the preceding behavior but it will be
reverted to the preceding behavior for now to avoid downstream impact.

The test fails in Turbine because it is stricter than Psi and throws
an exception because it cannot resolve the field.

Bug: 354633349
Test: ./gradlew
Change-Id: I27cda041a06853a1840d67fcadbe6c4afd5cb3d6