Do not require Kotlin positive ints to be prefixed with a +
Previously, when formatting an explicitly positive number for Kotlin,
e.g. `+17`, it would use the whole expression so include the unary
`+`. Supporting that in `DefaultMethodItem.legacyDefaultValue()` would
require additional work. However, it does not appear that AndroidX has
any such values so this change modifies the test and uses
`DefaultMethodItem.legacyDefaultValue()` for ints.

Bug: 354633349
Test: ./gradlew
      # Tested in Android and AndroidX and does not affect signature
      # files in either or stubs in the former.
Change-Id: I27b66fb1d900f6f517210a5c4a19c6c7e88ddd25