Include wasUnwrappedInSource in ArrayValue.snapshot(...)
Previously, `ArrayValue.snapshot(...)` did not copy
`wasUnwrappedInSource` to the snapshot which meant that code that
relied on that would break if it was run against the snapshot. Up until
now that was not an issue as there was no such code. However, a
follow-up change will make use of `MethodItem.defaultValue` instead of
`legacyDefaultValue()` in tests that rely on `wasUnwrappedInSource`.

This change corrects that.

Bug: 354633349
Test: ./gradlew
Change-Id: Icb53ae25bbd7bc3bce24e4cbcc80a9150e330f09