Stop creating DefaultAnnotationAttributes directly
Previously, `AnnotationsMerger` created `DefaultAnnotationAttribute`s
directly to pass into `Codebase.createAnnotationWithAttributes(...)`.
Those `DefaultAnnotationAttribute`s had to use `ValueProvider.UNKNOWN`
because they were created without an `AnnotationItem`. The
`DefaultAnnotationAttribute`s were used as if they were a pair of
attribute name and attribute value as a source string. The only real
benefit of that was to reuse `formatAnnotationItem`. However, that
was not worth creating special forms of `DefaultAnnotationAttribute`
whose `value` property could not be accessed.

This change replaces the `List<AnnotationAttribute>` with a
`List<Pair<String, String>>` and duplicates a simplified form of the
`formatAnnotationItem(...)` method. That allowed the removal of the
special `DefaultAnnotationAttribute.create(...)` methods as well as
`ValueProvider.UNKNOWN`.

This is preparation for removal of `legacyValue()` from
`AnnotationAttribute`.

Bug: 354633349
Test: ./gradlew
Change-Id: I34ba1ff6ecd9b8d713953f0eb3889c5b89109b73