Support formatting non-literal ints as hexadecimal
The `AnnotationItem.toSource()` formats non-literal ints using
hexadecimal instead of decimal. Where a non-literal int is one that was
not represented as a literal in the source but instead was computed
from an expression. Historically, Psi has considered a negative int to
be a unary minus expression, even when loaded from a jar where it is
represented as a negative int.

This change adds a `nonLiteralInSource` property to `DefaultIntValue`
to track whether it was a non-literal in the source (or negative) and
adds `ValueStringConfiguration.nonLiteralIntFormat` to control how a
non-literal int is formatted.

As with the previous

Bug: 399365344
Test: ./gradlew
Change-Id: I9c7f569ce938cda3cd63137855b665ac2917e203