Add ValueExamples for casts to byte/short
Psi treats values obtained from expressions differently to those
obtained from literals. In particular, it will format bytes and shorts
obtained from literals as decimals while those obtained from
expressions, e.g. `(byte) 123`, will be formatted as hexadecimal but
without a leading `0x`. That means that they will be parsed as decimal
and either fail (because `a-f` are not decimal digits), or, even worse,
succeed but read in the wrong number. e.g. `(byte) 116` will be written
out as `74` (in hex) but read in as `74` (in decimal).

This add tests for those cases.

Bug: 354633349
Test: ./gradlew
Change-Id: Idac2752ab7c91c7b3f7b36dc2d74b33abc9483e1