Always format bytes and shorts as decimals
Previously, byte and short values that were represented in the source
as literals were formatted as decimals and those that were computed
from an expression were formatted as hexadecimal. Unfortunately, their
format did not include a leading `0x` which meant they would not be
parsed as hex and would either fail to parse or be parsed as decimal
resultin in the wrong value being read.

This change forces them to be written out as decimal. It does that,
rather than just add the leading `0x` to make it consistent with how
they are written out when they are represented in the source as
literals.

This is a change in behavior that could impact signature files but as
the alternative is broken signature files this is an improvement. It is
unlikely that this will impact any signature files as those signature
files would probably result in failures, e.g. failing to parse, or
reporting compatibility errors due to reading the wrong value.

Bug: 354633349
Test: ./gradlew
Change-Id: Ic449dfd9f37155e6db85514d4554dc2c2b4c05a7