Allow ValueExample to support non-Java examples
Previously, there was no way to filter an example from being included
in the jar. That meant that passing say `""` to `javaType` and
`javaExpression` to indicate that java was not supported would result
in a compilation error when trying to build the jar. This change makes
the jar generation code obey the `validForInputFormats` property.

It also gives `javaType` and `javaExpression` a default value of `""`.

Bug: 416755106
Test: ./gradlew
Change-Id: I145b8553655c0dc0c5452bf1c00bc9a1522de97f