Generalize mapping of unsigned values to signed
Previously, Kotlin unsigned values were only mapped to the equivalent
Java signed if they were present as unsigned literals. This change
extracts it and uses it for the result of evaluation too.

This revealed an issue with the `ConstantEvaluator` not handling
unsigned values in K2. It works in K1 but only because they are
converted to Java signed values first. That required a couple of tests
to be baselined for K2 because they returned `null`. See b/417781943.

Bug: 354633349
Test: ./gradlew
Change-Id: Ib778c9854bdaeed8b11f8e7165e9bf6771dcece2