Store original value in DefaultNumericValue
Previously, `DefaultNumberValue` would store an indication of whether
the original value was an `int` or not as that was needed to maintain
the behavior of `MethodItem.legacyDefaultValue()`. However,
`AnnotationItem.toSource()` also needs to do something similar for
`float` so this change replaces `wasOriginallySpecifiedAsInt` with
`originalValue`.

Note, the goal with this legacy support is to do the minimal necessary
to support the existing uses, not maintain exact behavior across all
possible inputs even if those inputs are not actually used.

Bug: 354633349
Test: ./gradlew
Change-Id: Iaf62443fa4a8ab5d660a047ac179b361e1cf9a23