TurbineValueFactory: Preserve original value for field
Previously, `TurbineValueFactory` would use the built-in Turbine
support for getting the constant value of a field. The purpose was to
save time. However, it loses the original value which causes issues
when generating legacy formatting which relies on that.

This change makes Turbine use the same method as Psi to ensure
consistent behavior which fixes a couple of the value tests.

Bug: 354633349
Test: ./gradlew
Change-Id: I7c23c5e20dfceab1c6aaabb7c0dfe9e32358d904