PsiValueFactory: Determine if field is in companion object
Previously, `PsiValueFactory` ignored whether a field was in a
companion object and so the resulting `FieldReferenceValue`s did not
include `.Companion` in their qualified class name. That is required
for legacy `AnnotationItem.toSource()` behavior and so this change
tracks whether a field is in a Kotlin companion class or not.

It does that by relying on the fact that resolving the receiver in
a qualified field name like `Foo.FIELD` will resolve `Foo` to
`Foo.Companion` if the class has a companion object.

This adds a test to show the behavior of this with a companion object
with a non-standard name. A follow-up change will switch
`UAnnotationItem.toSource()` to use the `LegacyValueFormatter` which
will rely on this behavior.

Bug: 354633349
Test: ./gradlew
Change-Id: I7ae3ac7d876d7fcd27c8b4392a90e067000dd8b8