PsiValueFactory: Detect and deconstruct numeric conversion calls
Previously, something like `Foo.INT_FIELD.toLong()` would be
evaluated and replaced with its actual value. This change modifies
`PsiValueFactory` to match that pattern and deconstruct it to
create a `FieldReferenceValue` with `explicitConversionTo` set.

That causes two types of changes to the `ValueExample`s that test
that behavior:
1. `MethodItem.legacyDefaultValue()` changes from using the constant
   value to the field reference. While that behavior change could
   affect downstream users it does not appear as though it causes any
   issues for AndroidX.
2. The `expectedValue` changes from a primitive to a primtive wrapped
   in a field reference which is the intended effect of this change.
   This also fixes these tests in the text model.

This is needed for replacing `UAnnotationItem.toSource()` with
`LegacyValueFormatter.ANNOTATION_SOURCE_FORMATTER`.

Bug: 354633349
Test: ./gradlew
Change-Id: I940e74ec6b0ec2b83313d048b136a8505c98141a