Add tests to show misbehavior of Psi for @setparam:
Psi has a bug where it does not correctly resolve a `UAnnotation` to
its `PsiClass` (`UAnnotation.resolve()` returns `null`). That is
despite being able to do that for exactly the same annotation if it is
specified using `@get:` instead. That bug affects the creation of the
`Value` for its attributes as without the class
`ValueFactory.createLiteralValue(...)` cannot correctly convert the
value to the correct type.

This change adds tests to show that misbehavior which will be fixed in
a follow-up change. Note, it does not add tests for `UByte` and
`UShort` as they are not valid types for annotation attributes.

Bug: 354633349
Test: m checkapi
Change-Id: Iba6adbc8e1728f152583a6935f47e7bdc1114ae8