Work around Psi issue with @setparam: annotations
Change https://r.android.com/3640948 added a test to show the
misbehavior of `@setparam:` caused by `UAnnotationItem.resolve()`
always returning `null` instead of the annotation's `PsiClass`. That in
turn meant that the type of the annotation's attributes could not be
determined which resulted in `optionalTypeItem` being `null` for those
attribute's values. That resulted in a type being treated as `int`
(because it was specified using an `int` literal) instead of `long`.

This change works around that by using the fact that Psi has already
cast the `int` type to `byte`, `long` or `short` to determine the type
and then synthesizes a `TypeItem` to use instead of `optionalTypeItem`.

Bug: 354633349
Test: m checkapi
Change-Id: Ifdec882971a27fe3bb4ee0007bcded9d8507b1b7