ValueParser: Support conversion function call on field reference
Previously, `ValueParser` would not handle something like
`AClass.FIELD.toLong()` correctly. It would treat it as a Kotlin
`AnnotationValue` for an annotation class `AClass.FIELD.toLong`.

This change updates the pattern for matching fields to include an
optional conversion function call, e.g. `.toLong()` or `.toByte()`.
That fixes a number of the broken tests. The remainder will be
fixed in the following change.

Bug: 354633349
Test: ./gradlew
Change-Id: I01213a330aa6e9fe932bf068202266030c5b23c1