Another work around for a missing java.lang.String class
Since https://r.android.com/3616183 was submitted other code was added
which relies on resolving `String` to `java.lang.String` which breaks
on tradefed APIs. This change works around that issue. It also adds
tradefed as a target for the gather script so that any similar issues
will be detected early.

Bug: 415333758
Test: ./gradlew
      # Run the following before and after merging this change.
      # Before merging it will fail after merging it will work.
      m out/soong/.intermediates/tools/tradefederation/core/tradefed-doc-stubs/linux_glibc_common/everything/tradefed-doc-stubs-stubs.srcjar
Change-Id: Ica46238d4bc6c59871cbecd261afb12b1b830d0e