CommonAnnotationItemTest: Combine duplicate methods
Previously, `CommonAnnotationItemTest` had what were essentially
duplicate tests (same input, different checks), one for testing the
value and one for testing `toSource()`. The latter is not really
something that should be provided by or tested with the model. This
change combines the mostly duplicate tests in preparation for
removing the `toSource()` checks.

A couple of tests were slightly different:
1. The `char` tests. One used `b` another used `\uff00`. The latter
   was used in the combined method.
2. The `class` tests. One used fully qualified paths in signature
   but not Java, the other did not. The combined method used fully
   qualified in both to ensure the test passed but it made the test
   slightly unrealistic. It will be addressed in a follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: I127623270abd9733bc0254950124a264b8506eb3