Add test of how nextSdkVersion is currently set
Currently, `nextSdkVersion` in `VersionedSignatureApi` is set from the
`--current-version` (and `--current-codename` if set to something other
than `REL`). This change adds a standalone test to show that. The test
will make it easy to see the effect of a follow-up change which will
add a new command line option that can be used to set it directly.

Bug: 424440294
Test: ./gradlew
Change-Id: Ie8c616b8a53bd12a884c76393703d379e063d2f6