Add --check-compatibility enabled/disabled option
Previously, specifying one of the `--check-compatibility:...` options
would always cause compatibility checks to be performed against the
provided previously released API. That was an issue when the
`--check-compatibility:...` options were only used to provide the
previously released API to which flagged APIs can be reverted.

This change adds a `--check-compatibility` option that can disable the
compatibility checks for that use case.

Bug: 425309457
Test: ./gradlew
Change-Id: I25560668e044b1827b8363d9aa4bfcf0712dd92d