Expose PreviouslyReleasedApi from CompatibilityCheckOptions
Previously, `CompatibilityCheckOptions` provided the previously
released API through the `previouslyReleasedCodebase(...)` method which
returned an optional `Codebase`. That made it expensive to determine if
a previously released API had been provided as the only way to check
was to load the `Codebase`. This change exposes the
`PreviouslyReleasedApi` object instead which will allow a future change
to make sure that a previously released API was always provided if API
flags were provided.

Bug: 425309457
Test: ./gradlew
Change-Id: I556ec70d3a053f96b3f305b9bcc1dd2f28c5c56b