Disallow api flags without previously released API
Previously, it was possible to provide API flags via `--config-file`
without providing the previously released API to which any flagged APIs
would be reverted if their associated flag was disabled. That led to
lots of secondary issues (due to flagged APIs being removed instead of
reverted to the previously released API). They were time consuming to
triage.

This change fails early if API flags are provided without a previously
released API to which flagged APIs could be reverted.

Bug: 425309457
Test: ./gradlew
Change-Id: I32f6a436d6001584acfef930b8d5fb0cd813a95f