Treat future version as identical to latest version
Previously, when specifying `--api-version-for-sdk-extension` to be a
future version like `9999` it could result in APIs that were added in
the latest sources being treated as if they were removed. That is
because they were not added to the future version. This change treats
the future version to be identical to the latest version to avoid that.

Bug: 424440294
Test: ./gradlew
Change-Id: Ibab3642421d7970e45384d5364c8e9d2f28fe9c0