Fix incorrect type for bestStubConstructorComparator
Previously, it explicitly specified the type on the call to
`comparingInt(...)` as `ConstructorItem?`. However, it was not actually
nullable and most likely only worked because of Kotlin being lenient
when dealing with platform nullability. With Kotlin 2.2 that produces a
warning which fails the build due to `-Werror`.

This change corrects the type so it will work in Kotlin 2.2.

Bug: 427971908
Test: ./gradlew (before and after switching to Kotlin 2.2)
Change-Id: I0737dd19f35b56cc67d2de7d648326e17c1bf8a5