Catch changing a method from public to private with source input
When a method appears to be removed from a class, metalava double checks if there is a copy present by doing a search including super classes, to make sure the method was actually removed. However, it was doing this check twice: once in ComparisonVisitor.dispatchToRemovedOrCompareIfItemWasMoved, where the search only considers methods in the API, and again in CompatibilityCheck.removedCallableItem, where the search also considers methods that are not part of the API surface (private, hidden, etc). This meant that if a method became private, it was still being seen in CompatibilityCheck.removedCallableItem even though it should be considered removed from the API surface.

This removes the check in CompatibilityCheck.removedCallableItem, since ComparisonVisitor.dispatchToRemovedOrCompareIfItemWasMoved is called first and is already searching for a matching super method. The one case handled in CompatibilityCheck.removedCallableItem but not ComparisonVisitor.dispatchToRemovedOrCompareIfItemWasMoved is unstable reverted APIs when using flagged APIs. The handling of this case has been moved to ComparisonVisitor.dispatchToRemovedOrCompareIfItemWasMoved.

Bug: 433276725
Test: Added case in CompatibilityCheckTest, updated existing tests with incorrect visibility, flagged API tests still pass
Change-Id: Ie062439ae4e5276387ed2c8e6b689d2f5dfaa1f0