Update PsiEnvironmentManager close handling
Some tests open two PsiEnvironmentManagers, which means that when they get closed only the last one to close can assert that everything has been disposed of. (Added in aosp/3076965)

The `KotlinCoreEnvironment.applicationEnvironment` property which is currently used to determine which PsiEnvironmentManager can assert that everything has been disposed is being deprecated. This replaces it with a count of open enviroment managers, which is incremented when a manager is constructed and decremented when a manager is closed, so that only one manager will declare that evertying is closed.

Test: CommonModelTest
Bug: 434743977
Change-Id: I80ad33f8fcaa574b6066ebc25430ed618808f3ed