Avoid using String.toItemDocumentationFactory() in metalava-model
Previously, the `String.toItemDocumentationFactory()` extension
function was only used in source models (i.e. turbine and psi) apart
from one use in `CodebaseAssembler.createPackageItem(...)`. This
change removes the latter use by delegating the work to subclasses
of `CodebaseAssembler`.

This is preparation for moving `DefaultItemDocumentation` (which is
indirectly used by `String.toItemDocumentationFactory()` into
`metalava-model-source` so it can make use of `JavadocParser`.

Bug: 429965593
Test: ./gradlew
Change-Id: I330f9535752679232cd49a7927ef03aaaed48d49