Add DocComment model and parser
The first use that is made of the Javadoc/KDoc comment in Metalava is
to search for `@hide`, `@doconly` and `@removed` doc tags to set the
`isHidden`, `isDocOnly` and `isRemoved` properties in
`ItemDocumentation`. Those checks are performed on most classes and
class members in the source, not just the ones that are part of the API
surface, so they have to be quick but fortunately, they only require
quite simple parsing of the comment.

This change adds `DocComment` model, a `DocCommentParser` for
constructing a `DocComment` from some text and some tests to show
that it behaves correctly. This initial implementation supports the
standard Javadoc/KDoc syntax. A follow-up change will customize it to
handle some Android special cases.

Bug: 429965593
Test: ./gradlew
Change-Id: I3eae984041f290c2929ba1244075b941e7530a1d