Update compatibility checks to account for differences in target languages
If two callables have the same signature but disjoint target language sets, it makes sense to treat them as distinct elements when performing compatibility checks. That way, if there is another element with the same signature and different target languages, the right elements will be compared.

Additionally, when checking if a method has had a target language removed, this adds a check that there isn't a different overload which works for the removed target language. This is useful in the case that one method entry is split into two, with different kotlin and bytecode entries in the new signature.

Bug: 445490139
Test: CompatibilityCheckTest

Change-Id: Icfbb79c3e902acdf6a6c08da785c5abbbd95cbcd