Prepare Baseline to support conditionally suppressing element ids
Currently, whenever an element id is added to the `Baseline` that will
always suppress any other messages that have the same element id. That
means that if messages with the same element id but different locations
are reported only the first will be output. That makes it tedious and
time consuming to fix issues based on the error messages as only one is
reported at a time.

This change prepares `Baseline` to differentiate between element ids
loaded from a baseline file (which should suppress the message) and
element ids added when reporting a message which should not suppress
future messages with the same element id.

Bug: 429965593
Test: ./gradlew
Change-Id: I6fdf0be9a271379de7c181ed3671fafb2711873b