Only suppress messages for element ids read from a baseline file
Previously, reporting a message would suppress all subsequent messages
that had the same element id even if they were at different locations.
That meant that if multiple messages with the same element id but
different locations were reported then only the first would be output.
This change will only suppress messages for element ids that are read
from a baseline file. Element ids that are added to the `Baseline`
because a message is being reported will not suppress subsequent
messages with the same element id.

This fixes an issue with a `JavadocTest` where multiple messages
were being reported with the same element id but only the first was
being output.

Bug: 429965593
Test: ./gradlew
Change-Id: I9aa96f6bbd678450213a4b4d893d7c45b706b80a