Make JavadocParser.parse(...) parse a String range
Previously, it would parse the whole `text` `String`. This change makes
it parse a substring. Preparation for follow-up work that will change
`JavadocParser` to parse a description instead of the whole comment.
The `DocCommentParser` already handles splitting it into a main
description and separate block tags each with their own description.

Bug: 429965593
Test: ./gradlew
Change-Id: I1c99ec767017a42c3209c5e96f4f7277ed5edad4