Trim leading newlines from javadoc comment
Previously, leading newlines at the beginning of a javadoc comment were
preserved in `JavadocContent` objects. e.g. after removing the javadoc
start token ("/**") and the " *" indentation there is a newline before
"Blah" in the following:
```
/**
 * Blah
 */
```

But there is not a newline in the following:
```
/** Blah */
```

The newline makes it difficult to consistently format Javadoc comments
as the above two contain the same textual content but would be
formatted differently.

This change removes leading newlines at the start of a Javadoc comment
as part of the work to normalize Javadoc comments for consistent
formatting.

Bug: 429965593
Test: ./gradlew
Change-Id: I68c16bdc9ac45ce37343e8957ba1738af617f93d