Make JavadocParser treat */ within a comment as an error
Previously, it was just treated as normal text. This change fixes that
by adding it as a the `COMMENT_END` token that is not actually used in
the grammar. It was also necessary to add EOF to the end of the
`description` rule otherwise the parser would just stop when it see the
`COMMENT_END` token.

This caused the error tests added in the previous change to fail with
an exception. That also revealed an issue with the line number and
character positions reported for block tag description error.

Bug: 429965593
Test: ./gradlew
Change-Id: Icf4782caf77f75715cd27aab972f67c491bd47ad