Remove unnecessary blank lines from stub files
Preparation for improving the testing of stub files in `DriverTest` by
not removing blank lines from the stub file before verifying its
contents. This removes unnecessary blank lines to reduce the number of
tests that will need updating when blank lines are kept.

This also reduces the size of the stub files but does make them
slightly less readable. That trade off is worth it because people do
not tend to read stub files.

Bug: 448617351
Test: ./gradlew
Change-Id: I5a666b8a2c468391782c88d330008147f81f5ea2