Add leading asterisk where possible
Previously, comment lines that did not start with a leading asterisk
did not have one added. This change adds " *" to all lines that do not
have a leading asterisk and which do not start with "/". That ensures
consistency with follow-up changes that will print Javadoc comments
from the `JavadocContent` model which does not track which of the
comment lines originally had a leading asterisk.

Bug: 448617351
Test: ./gradlew
Change-Id: Ibae77ae63ac76b9f9535f4796d7cf73e38e0c5d3