Add/change test to show whitespace before */ is not trimmed
The `DocCommentParser` minimizes the amount of work it does as it is
used on most of the sources, even those parts that do not end up in the
API (e.g. to find `@hide`). So, it intentionally does not trim trailing
whitespace from the ends of the `DefaultDocDescription`s it creates.
The `DefaultDocDescription.toString()` (called from some tests) does
remove it but it is not removed when parsing. That can lead to the
content being included in unclosed inline tags because inline tags do
not remove trailing whitespace as that could be significant, e.g. if
they are followed by other text.

This change adds a test/changes an existing test to illustrate the
broken behavior.

Bug: 429965593
Test: ./gradlew
Change-Id: If3430ee4280a7704fab98ca67abf9b0747790d77