Move line/char correction to DefaultDocDescription
Previously, `JavadocErrorListener` would correct the `lineOffset` and
`charOffset` for an error to make it relative to the start of the
`DefaultDocDescription`'s `text` property. That meant it would not be
applied if `JavadocParser.parse(...)` reported an issue directly to the
`reporter`. That will be required in a follow-up change.

This change moves the correction logic to `DefaultDocDescription` by
making it implement `DocumentationIssueReporter` and passing itself to
`JavadocParser.parse(...)`. Its implementation applies the correction
and then forwards it to the `reporter` with which it was constructed.

Bug: 450021237
Test: ./gradlew
Change-Id: I6b9f3e6d4e0cc49cc512332d278a96f3a4c468fc