Stop removeDeprecatedSection(String) from removing empty comments
Previously, `removeDeprecatedSection(String)` would remove the whole
comment if after removing the `@deprecated` section the comment was
empty. That had two issues:
1. A blank comment and no comment can behave differently. e.g. for
   overloaded methods it can affect what documentation is inherited.
   This is especially true when it is processed by `doclava` whose
   behavior is not well specified but most likely implements an
   older version of Javadoc behavior which did not automatically
   inherit comments.

2. It is not particularly realistic test. An API should not just have
   a single `@deprecated` section, it should have some content in the
   main description.

Preparation for switching to perform this removal on the `DocComment`
model which does differentiate between no comment and an empty
comment.

The `@Suppress("DeprecatedIsStillUsed")` as its previous position
prevented the IDE from syntax highlighting the following string as Java
code.

Bug: 450228132
Test: ./gradlew
Change-Id: I580f202c9080e412ac9fa8c1ff37c421552f5251