Override existing apiSince and deprecatedSince
Previously, if the source already contained an `apiSince` or
`deprecatedSince` block tag an error would be reported but the
`apiSince` would not be changed. This change makes it so that any
existing block tags are removed before the new one is added. That makes
them consistent with how `sdkExtSince` is already handled.

Bug: 450228132
Test: ./gradlew
Change-Id: Ia0ba51211f1ab2029ba6b6ed36aef55bbc2dc122