Add DocContent/Owner abstraction
Currently, `DocAnalyzer` uses `ItemDocumentation` methods
`findMainDocumentation(...)` and `findTagDocumentation(...)` to
retrieve parts of a Javadoc comment for appending to other
documentation or searching for the word `null`. They return `String`s
which means it is not possible for them to return instances of the
`JavadocContent` or `DocDescription` model.

This change adds the `DocContent` abstraction to replace the use of
`String`. It also adds the `DocContentOwner` abstraction around the
`DocContent` that will be used in future to allow changes to the
`DocContent`. It also adds the `mainDescription` property and
`blockTagDescription(String)` method to `ItemDocumentation` to
provide access to them both and replace the existing `find*(...)`
methods mentioned above.

Bug: 450228132
Test: ./gradlew
Change-Id: Ieaf5314423ab972548b87472c94cb47608e33a8f