Treat @exception as an alias for @throws
Previously, `@exception` and `@throws` were treated as separate block
tags. However, they are basically the same and while `@exception` was
added first `@throws` is the recommended tag to use. This change treats
`@exception` as an alias for `@throws` so any occurrence of
`@exception` in the source will be replaced with `@throws` in the
stubs.

Bug: 452699372
Test: ./gradlew
Change-Id: I534fa7cd8abc2c69df57b7fd506b0e5ddab80d62