Document and enforce JavadocContent constraints
Previously, the constraints that `JavadocContentList` must contain at
least two `JavadocContent` instances and that `JavadocText` cannot be
empty were implicit in the code that uses them, e.g.
* An empty `JavadocContentList` or empty `JavadocText` was never
  created and instead were replaced with a `null` `JavadocContent`.
* A `JavadocContentList` that only contained a single item was
  replaced with its single `JavadocContent`.

This change enforces those constraints in the classes themselves. This
is in preparation for future changes that will add more code to
manipulate `JavadocContent` instances to protect against them from
breaking those constraints which are relied upon elsewhere.

Bug: 450228132
Test: ./gradlew
Change-Id: Ifa4893990e6d68616556a46008968ea1c75d1e2a