Remove mutationListener from DefaultDocComment
Previously, `DefaultDocComment` took both a `DocCommentContext` and a
`DocCommentMutationListener` but the latter can be retrieved from the
former. This change removes the `DocCommentMutationListener` and just
gets it from `DocCommentContext` when needed.

This might seem like a pointless change but it removes some work from
the creation of `DefaultDocComment` (which is performed for almost
every Javadoc comment in the sources) to when it is modified which only
happens for comments that are part of the API surface.

Bug: 450228132
Test: ./gradlew
Change-Id: I72d63d687b8490ba7cf438e2bd0809c70d1bbb80