Simplify handleKotlinDeprecation(...)
Previously, it would add the text from the Kotlin `Deprecated`
annotation to the `@deprecated` Javadoc tag (creating it if it did not
exist) if the annotation provided an explanation message and the
documentation did not already contain the text.

However, Metalava does not generate Kotlin stubs, or generate Java
stubs from Kotlin code, so the actual content is irrelevant. After some
investigation it turns out that the reason it adds the `@deprecated`
Javadoc tag is just to avoid reporting missing `@deprecated` Javadoc
tags on items which are deprecated using an annotation.

This change maintains that behavior by replacing the check to see if
the description contains the `Deprecated` message with one that just
checks to see if the `@deprecated` block tag is present. That will make
it easier to migrate to using the model.

Bug: 450228132
Test: ./gradlew
Change-Id: I28a05272d2e5691a5002e6d19e83e1c355de56e8