Prepare DocContent for appending to other documentation
Currently, some `DocAnalyzer` takes the text content from annotation
documentation block tags (e.g. `@memberDoc`) and inserts them into the
documentation of some other items. Before it does that it trims leading
whitespace and fully qualifies any source references. The latter is
needed as the source references will be relative to the annotation and
not where it is appended.

Note: There is a slight issue with the fully qualification that is
    currently done as that combines resolving the references to a
    specific `Item` and then using relative references where possible,
    e.g. to a member of the same class, to a class in the same package.
    Those relative references do not work when the content is placed in
    a different class or package. That will be fixed in future as the
    resolving and simplifying will be done separately.

This change adds similar support for `DocContent`. It is specialized
for this use but longer term this will not be needed as whitespace will
be normalized on input and formatted correctly on output. Similarly,
source references will be resolved on input and simplified on output.

Bug: 450228132
Test: ./gradlew
Change-Id: I41ff2de46e98e99a5f81fe60b7d8f94a7498339a