Use DocContent model in AndroidApiChecks.checkVariable(...)
Previously, it used custom parsing code `getDocumentation(...)` to get
the main description, `@return` description or `@param` descriptions.
That was expensive and so it would cache the last use.

This change switches it to use the `DocContent` model which is much
simpler, faster and does not require caching. It also improved the
documentation and made some minor improvements to what is now legacy
code in the way it handles types.

Bug: 450228132
Test: ./gradlew
Change-Id: I036207520fb30c663ad3fae502ac826246a684e3