Add '.' to terminate sentence fragment before appending
Previously, appending content to existing content would insert a `<br>`
between the existing and new content. This change adds a check to see
if the existing content was a sentence fragment (i.e. did not have a
period at the end) and if it is then it will add a period at the end of
the existing content and before the `<br>`.

Bug: 452926384
Test: ./gradlew
Change-Id: Ie9e05565d4837255d954d4595470c538b426a203