Delegate SourceFileSnapshot.getImports() to original SourceFile
Previously, `SourceFileSnapshot` did not implement `getImports()` which
meant it use the default implementation of `getImports { true }`. That
delegated to `FilteringSourceFile.getImports {...}` which delegates to
the filtered `SourceFile` without applying the filters. That meant some
imports being kept that should have been removed.

This change delegates `SourceFileSnapshot.getImports()` to the original
`SourceFile` which applies the filters correctly and removes the
invalid import. It also makes `FilteringSourceFile.getImports { ... }`
apply the filtering to avoid similar issues in future.

Bug: 447588621
Test: ./gradlew
Change-Id: I1290e05bb3f54ca734d0f4dc5810034f8c5e2a1f