Add test to show misbehavior of SourceFile snapshot
Change https://r.android.com/3819385 added the snapshotting of
`SourceFile`s but unfortunately, it was slightly broken in two ways
when combined with `FilteringSourceFile`.

Firstly, `FilteringSourceFile.snapshot(Codebase)` was delegated to the
`SourceFile` that it was filtering which meant the snapshot did not
apply the filtering to the `getImports()` method.

Secondly, `SourceFileSnapshot.getImports()` was not delegated to the
original `SourceFile` so it ended up calling `getImports { true }`
which called `FilteringSourceFile.getImports { true }` which also
bypassed the filtering.

This change adds a test to show that misbehavior.

Bug: 447588621
Test: ./gradlew
Change-Id: I39d84c4575b6ea6762f787b8996a7408dfe84f35