Add tag type specific printing
Previously, `TagData.printAfterTagType(...)` allowed some tag type
specific formatting but it was limited to prefixing the content with
information from the `TagData`, e.g. the parameter name. However,
`{@link}` tags need to have more control. In particular it needs to
know if there is any additional content (the label) as that will affect
whether it will generate a label or not.

This change replaces `printAfterTagType(...)` with
`printTagContents(...)` and `DocTag.printTagContents(...)` delegates
the printing to it if available, otherwise falling back to just
printing `DocTag.content`.

Bug: 447588621
Test: ./gradlew
Change-Id: I90a0f809df4045797cb5346c9ec5011553812741