Show that invalid documentation issues can be reported
Currently, misusing `@hide` is detected during parsing a `DocComment`
that is done once before the documentation is first mutated and then
maybe once when printing (if fully qualifying changes the text). If
the position of the `@hide` is changed, e.g. because mutating changed
it from a single line comment to a multi-line comment, then that can
result in multiple issues being reported for the same problem.

This change adds a test that shows that behavior. A follow-up change
will fix it.

Bug: 447588621
Test: ./gradlew
Change-Id: I57a0fc30f1a4b7335f5d200994dd6864ceea7874