Stop ignoring empty doc string in itemDocumentationFactoryForDecl(...)
Psi returns an empty doc string for a non-existent Javadoc comment.
However, Turbine returns a `null` doc string in that case. Also,
Turbine currently removes the javadoc start token from the front (but
not the end token) so the String is never empty. Both of these make the
empty String check unnecessary.

However, in future Turbine may also prune the end javadoc token which
would mean that an empty comment would return an empty String. In that
case the empty String check here would be incorrect as it would lose
any position information about the Javadoc comment.

So, this change removes the check.

Bug: 458362596
Test: ./gradlew
Change-Id: I39fba7f0eae14311d1118acb6916abe66c7a3d04