Add basic conditional Javadoc processing
Adds basic support for parsing and processing conditional javadoc. e.g.
something like:
```
  /**
   * .....
   * {@if (flag(Flag.NEW_API)) {Describe how new api impacts this.}
   */
```

Or:

```
  /**
   * .....
   * {@if (flag(Flag.CHANGED_API)) {
   *     Describe new behavior.
   *   } else {
   *     Describe old behavior.
   *   }
   * }
   */
```

The expressions are evaluated while parsing the Javadoc comment so the
rest of Metalava will see a comment based on the flag status. That
could result in a discrepancy with the documented item if that does not
also reflect the flag status. In that case a check that compares the
two could result in false positives. Those issues will be addressed in
future changes.

Bug: 429965593
Test: ./gradlew
Change-Id: I1d885b0e05b2788aa4b94358b300657326dcdf8a