Integrate conditional processing into ItemDocumentation
Previously, it was only supported internally in tests. This change adds
support for using conditional Javadoc in `ItemDocumentation` making it
usable in actual source code.

The implementation does not handle all possible ways to specify the
flag field (TODOs added) but it does support the most common ways of
using the automatically generated flag fields.

Bug: 429965593
Test: ./gradlew
Change-Id: Iecd660d22b1e8f9754b74f6595461fe496503319