Remove COMMENT_END and associated tests
It is pointless for the `JavadocParser` to detect a javadoc end of
comment token in the body of the comment as that will never happen in
real code. That is because the `SourceParser` implementation will have
already truncated the javadoc comment at the end of comment token. If
that was followed by additional text then that would have most likely
caused syntax errors.

This change removes `COMMENT_END` and the tests that verify it caused a
parser error.

Bug: 448905711
Test: ./gradlew
Change-Id: I37a50085761ce5773b3b04c52473b258841873e4