Use a new lexer mode for the inline tag name
Previously, the name of an inline tag was matched by the `NAME` lexer
rule. To prevent it being matched by the `TEXT_CONTENT` rule the
characters used in `NAME` had to be excluded from `TEXT_CONTENT` and
then anywhere that `TEXT_CONTENT` was used `NAME` also had to be
allowed. That complicated the grammar and quickly becomes untenable
when other rules are added.

This change switches it to use a new lexer mode for the inline tag
name, using `pushMode(INLINE_TAG_MODE)`. A mode effectively creates a
sub-lexer that is completely independent of the default lexer. The
sub-lexer is only used to process the `INLINE_TAG_NAME` rule and that
switches back to the default lexer using `popMode`.

This revealed a minor issue in ANTLR's handling of `expectedTokens`
which this works around. It also slightly changed the error message
reported when there was a space between `{@` and the `INLINE_TAG_NAME`.

Although this might make the lexer slightly more complicated, by using
modes, it is preparation for upcoming work to add a conditional
`{@if ...}` tag which will make much more use of modes so this will
ensure consistency for them.

Bug: 429965593
Test: ./gradlew
Change-Id: Ib9b92061fae01e39a1e54312ec8e90bb5ca8c85b