Filter packages based on package.html correctly in Turbine
Previously, Turbine passed `{ true }` as the `packageNameFilter` to
`gatherPackageJavadoc(...)`. That meant it was necessary for the code
that processed the information returned from there to filter out
packages. That differed from Psi which supplied a proper filter that
checked for the existence of an actual package.

This change switches Turbine to use `isValidPackage(String)` to ensure
consistent behavior with Psi. That is needed as following changes will
refactor the package handling code.

Bug: 461491536
Test: ./gradlew
Change-Id: I7e6c951eeae8558fe2f2bcf3d91277b0fd48c96c