Remove @hide and @doconly block tags from stubs
Previously, Metalava would leave `@hide`, and `@doconly` in the stubs.
When those stubs were processed by `doclava` it would detect them and
possibly do something undesireable. e.g. if it saw an `@hide` then it
would ignore the API and not include it in the documentation.

The only case where Metalava would include an API with an `@hide` block
tag in the stubs was if that API was annotated with a show annotation,
e.g. `@SystemApi`. In that case Metalava removed the `@SystemApi`
annotation but kept the `@hide` which caused `doclava` to exclude the
documentation for all system APIs.

This change stops Metalava from including `@hide` and `@doconly` block
tags to prevent `doclava` ignoring system APIs.

Bug: 462333324
Test: ./gradlew
Change-Id: I471b4e948f11a408b86a36570c933f86706f08e5