Add extension functions to create duplicate and snapshot factories
Previously, code that needs an `ItemDocumentationFactory` that created
a duplicate, or took a snapshot of an `ItemDocumentation` would use a
method reference directly. This change adds methods to do that. This is
preparation for when `SelectableItem.documentation` can be null when
creating the factory is more complicated. It will make that change
simpler by only requiring a change to the extension functions not each
use.

Bug: 462419767
Test: ./gradlew
Change-Id: I850bfe37aa7d3d579a5899a77d8439eb4704a1bf