Add optionalDocumentation and requiredDocumentation
In order to migrate `documentation` from `ItemDocumentation` to
`ItemDocumentation?` and remove `NoItemDocumentation` it is helpful to
differentiate between those uses requiring a real `ItemDocumentation`
(i.e. not `NoItemDocumentation`) and those uses for which the
`ItemDocumentation` is optional and so can use `null` instead of
`NoItemDocumentation`.

This change adds two new abstract properties that will differ in
behavior when `documentation` is `NoItemDocumentation`:
* `optionalDocumentation` will return `null`.
* `requiredDocumentation` will throw an exception.

Existing uses of `documentation` in `metalava-model` and
`metalava-model-source` are updated to use one of the above new
properties where necessary. That is determined by what the code would
do if `documentation` is a `NoItemDocumentation`.
* If it always throws an error then use `requiredDocumentation`.
* If it always returns returns `null` then use `optionalDocumentation`.
* If it is an extension method that can be made to work on
  `ItemDocumentation?` then leave it.

Follow-up changes will apply those same rules for other uses of
`documentation`.

Once everything has been updated then `documentation` and
`optionalDocumentation` will be merged back into a single property
called `documentation` but with a type of `ItemDocumentation?`.

Bug: 462419767
Test: ./gradlew
Change-Id: Id2527c42a7dcbc7400aef38ece8a32336bdf78dd