Add ClassLoaderBasedCodebaseAssembler
Previously, `ClassLoaderBasedClassPathResolver` would use a
`TextCodebaseAssembler` and implement `resolveClass(...)` directly.
However, that duplicated functionality from the assembler. It could
not use the assembler directly as the `TextCodebaseAssembler` does not
support the behavior it needs. It was also inconsistent with how
Psi works.

This change moves the assembler like functionality out of
`ClassLoaderBasedClassPathResolver` and into a new
`ClassLoaderBasedCodebaseAssembler`. Removing duplication and making
it more consistent with the other models.

Preparation for adding support for resolving packages.

Bug: 462658629
Test: ./gradlew
Change-Id: I81f643de423ee964c992d67b17d30c4da7adca47