Add SourcePackageInfo to support merging with PackageDoc
Previously, `SourceCodebaseAssembler` would use `PackageInfo` for
the information returned from its subclasses, for merging with
`PackageDoc` and for returning to `PackageTracker`. The problem
with that was that merging with `PackageDoc` required the
`commentFactory` to be nullable but the `PackageTracker` would really
prefer it to not be nullable.

This change adds a `SourcePackageInfo` which is used for merging with
`PackageDoc`. At the moment it is identical to `PackageInfo` but a
follow-up change will change `PackageInfo` to stop `commentFactory`
being nullable.

Bug: 461491536
Test: ./gradlew
Change-Id: Idcc90b95ee117c2101e86c74724a9d326fab74ba