Allow getPackageInfoFromSource(...) to return `null`
Previously, implementations would return `PackageInfo.EMPTY` when they
could not find any information about the package. However that has a
`null` `commentFactory` property so could never be returned directly
and it is not clear when no information could be provided.

This change replaces that will `null` making it more obvious allowing
some optimizations in the `SourceCodebaseAssembler`.

Bug: 461491536
Test: ./gradlew
Change-Id: I4a47057d94899f649f3facf45c29d952a376e4ed