Remove global fastPathCheckResult
Previously, `fastPathCheckResult` was used to support testing the fast
path compatibility check that does not perform compatibility checking
if the generated signature file is byte-for-byte identical to the
previously released signature file. At the time the
`postAnalysisChecker` mechanism did not exist.

This change removes the global variable and uses the
`postAnalysisChecker` to get access to `CheckRequest` instances in
which the code stores the state of the fast path check.

Bug: 467355045
Test: ./gradlew
Change-Id: I4c342e093ecb480a04fd037c4fe3965ed45d08c5